Change cert-manager resolver secret data type
The cert-manager logs are throwing an error about

"error instantiating google clouddns challenge solver: Unable
to acquire config: invalid character 'e' looking for beginning of
value" "key"="networking/"

When reviewing how I had this previously setup and looking at the
example from cert-manager docs, the secret used for the clouddns
resolver key was stored with data instead of stringData, which affects
how the secret is stored, base64 vs not.
My hope is that changing this to match previous configs will get rid of
that error in cert-manager.