adjust traefik service ip to original from nginx
In hindsight this wasn't neccessary, but i gave traefik a different ip
than the one used for nginx originally thinking that they would conflict
if i just fed the existing ip to traefik while flux and helm were trying
to shuffle out nginx and deploy helm. In hindsight this wasn't
neccessary since this all works on eventual consistency, and a failure
to deploy traefik due to the conflict in address would have just meant
it was fixed on the next pass. Ohes well, fixing it now so dns matches
the right ip.