reorganize rook-ceph-cluster value indentation
The settings, particularly around select storage devices was setup
incorrectly. This meant it was just consuming all the media available to
it. Which, at the moment is fine, but in the future could be a problem.
The hope is that the changes here won't trigger a recreation of the
entire cluster and block pool. But i'll restore the volumes that's the
end result.