reduce cpu requests for media pods
the new rook-ceph upgrade has greatly increaed pod cpu requests and as a result, nodes are overprovisioned with cpu allotments, forcing other pods like most of the media ones to be unscheduled due to insufficient cpu on the nodes. Hoping this reduction in request will make room.

The nodes themselves are not highly cpu taxed at the moment.