remove transmission-downloads-pvc and backups
As it turns out, the reason volsync refused to back up any data from this volume is that the volume was not in use and was completely empty. I had switched this volume over to the nfs directly because i couldn't get readwritemany to work for the volume with multiple pods needing to write to the data at the same time. The downloads-pvc was just never removed and i forgot about it. Doh. No need to back this up, as it's backed up by the downloads-backup cron job.